Moved KordiRouter to core/navigation folder, created parts for routes to separate them in different files
