+++ lib/collections/pages/create_collection_first_step.dart
@@ -1,0 +1,10 @@
+part of 'package:kordi_mobile/collections/pages/create_collection_page.dart';
+
+class CreateCollectionFirstStep extends StatelessWidget {
+  const CreateCollectionFirstStep({super.key});
+
+  @override
+  Widget build(BuildContext context) {
+    return Container();
+  }
+}
