+++ lib/collections/pages/create_collection_second_step.dart
@@ -1,0 +1,10 @@
+part of 'package:kordi_mobile/collections/pages/create_collection_page.dart';
+
+class CreateCollectionSecondStep extends StatelessWidget {
+  const CreateCollectionSecondStep({super.key});
+
+  @override
+  Widget build(BuildContext context) {
+    return Container();
+  }
+}
