+++ lib/collections/pages/create_collection_third_step.dart
@@ -1,0 +1,10 @@
+part of 'package:kordi_mobile/collections/pages/create_collection_page.dart';
+
+class CreateCollectionThirdStep extends StatelessWidget {
+  const CreateCollectionThirdStep({super.key});
+
+  @override
+  Widget build(BuildContext context) {
+    return Container();
+  }
+}
