+++ lib/core/pages/kordi_scaffold.dart
@@ -1,8 +1,9 @@
 import 'package:flutter/material.dart';
 import 'package:flutter_bloc/flutter_bloc.dart';
+import 'package:go_router/go_router.dart';
 import 'package:kordi_mobile/collections/controllers/collections_filter/collections_filter_bloc.dart';
 import 'package:kordi_mobile/collections/controllers/get_collections/get_collections_cubit.dart';
-import 'package:kordi_mobile/core/utils/kordi_routes.dart';
+import 'package:kordi_mobile/core/navigation/kordi_router.dart';
 import 'package:kordi_mobile/core/widgets/kordi_drawer.dart';
 import 'package:kordi_mobile/core/widgets/kordi_search_button.dart';
 import 'package:kordi_mobile/dependency_injection.dart';
@@ -68,6 +69,19 @@ class KordiScaffold extends StatelessWidget {
           ],
         ),
         drawer: KordiDrawer(),
+        floatingActionButton: Builder(
+          builder: (context) {
+            if (GoRouterState.of(context).fullPath == '/collection') {
+              return FloatingActionButton(
+                child: Icon(Icons.add),
+                onPressed: () {
+                  // CreateCollectionPageRoute().go(context);
+                },
+              );
+            }
+            return const SizedBox();
+          },
+        ),
         body: child,
       ),
     );
