+++ lib/core/widgets/kordi_search_button.dart
@@ -1,6 +1,6 @@
 import 'package:flutter/material.dart';
 import 'package:go_router/go_router.dart';
-import 'package:kordi_mobile/core/utils/kordi_routes.dart';
+import 'package:kordi_mobile/core/navigation/kordi_router.dart';
 import 'package:kordi_mobile/core/widgets/collections_filter_dialog.dart';
 
 class KordiSearchButton extends StatelessWidget {
